Introduces Graph with vector<unique_ptr> collection of edges and nodes.
Also separates pure graph concept from the data associated with edges and nodes
for various purposes.

The reason for that is that many data associated with nodes and edges is only needed for a particular computation. After the computation they can be deleted. If they were all in the Node or Edge or their subclasses they would keep using memory.